% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnv_utils.R
\name{check_passport_columns}
\alias{check_passport_columns}
\title{Check and Standardize Passport Columns}
\usage{
check_passport_columns(passport_df)
}
\arguments{
\item{passport_df}{A dataframe containing passport information.}
}
\value{
The passport dataframe with standardized column names.
}
\description{
This function validates the column names in the passport dataframe, ensuring that all required columns
are present. It allows for variations in column naming conventions and standardizes the column names
to their primary forms.
}
\examples{
passport_df <- data.frame(Sample = c("S1", "S2"), generation = c("Parent", "Child"), Mother = c("M1", "M2"), fam = c("1", "2"))
standardized_df <- check_passport_columns(passport_df)

}
